Extract mergeProperties
