+++ lib/schema3Resolver.js
@@ -1,7 +1,7 @@
 "use strict";
 var fs = require('fs');
 var defined = require('./defined');
-var defaultValue = require('./defaultValue');
+var mergeProperties = require('./mergeProperties');
 var clone = require('./clone');
 var replaceRef = require('./replaceRef');
 
@@ -72,9 +72,14 @@ function extend(derived) {
     var base = derived['extends'];
     if (defined(base)) {
         delete derived['extends'];
-        // TODO: extends could be an array
-        mergeProperties(derived, base);
-
+        
+        if (Array.isArray(base)) {
+            for (var singleBase in base) {
+                mergeProperties(derived, base[singleBase]);
+            }
+        } else  {
+            mergeProperties(derived, base);
+        }
         extend(derived);
     }
 
@@ -87,35 +92,3 @@ function extend(derived) {
     }
 }
 
-/**
-* @function mergeProperties
-* Recusively takes properties within a schema reference ("the base") and merges the contents of
-* those properties into the derived schema.
-* @param  {object} derived - The schema that contains a reference to the 'base' schema.
-* @param  {object} base - The schema that was being referenced by 'derived'.
-* @return {object} The merged schema with the 'base' schema reference removed since the contents
-* have been merged into 'derived'.
-*/
-function mergeProperties(derived, base) {
-    for (var name in base) {
-        if (base.hasOwnProperty(name)) {
-            var baseProperty = base[name];
-
-            // Inherit from the base schema.  The derived schema overrides if it has the same property.
-            if (typeof baseProperty === 'object') {
-                derived[name] = defaultValue(derived[name], {});
-                var derivedProperty = derived[name];
-
-                for (var n in baseProperty) {
-                    if (baseProperty.hasOwnProperty(n)) {
-                        if (!defined(derivedProperty[n])) {
-                            derivedProperty[n] = clone(baseProperty[n], true);
-                        }
-                    }
-                }
-            } else if (!defined(derived[name])) {
-                derived[name] = clone(baseProperty, true);
-            }
-        }
-    }
-}
